requirements.txt
aio-pika==9.4.0
aiormq==6.8.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
pytest
click==8.1.7
fastapi==0.110.1
h11==0.14.0
httpcore==0.16.3
httptools==0.6.1
httpx==0.23.3
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
pamqp==3.3.0
pika==1.3.2
pluggy==1.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pytest==8.1.1
pytest-mock==3.14.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.37.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.20.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
